Add SignUpForm and auth_views

+++ products/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
+
 from django.utils import timezone
 from .models import Product
 from .forms import AddProductForm
@@ -25,33 +27,43 @@ def product_details(request, pk):
 
 
 def product_add(request):
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES)
+    if request.user.is_authenticated and request.user.is_superuser:
+        if request.method == 'POST':
+            form = AddProductForm(request.POST, request.FILES)
 
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-successful.html')
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-successful.html')
+        else:
+            form = AddProductForm()
+
+        return render(request, 'products/product-add.html', {'form': form})
     else:
-        form = AddProductForm()
-
-    return render(request, 'products/product-add.html', {'form': form})
+        return redirect('home')
 
 
 def product_edit(request, pk):
-    product = get_object_or_404(Product, pk=pk)
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES, instance=product)
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = get_object_or_404(Product, pk=pk)
+        if request.method == 'POST':
+            form = AddProductForm(
+                request.POST, request.FILES, instance=product)
 
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-successful.html')
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-successful.html')
+        else:
+            form = AddProductForm(instance=product)
+
+        return render(request, 'products/product-add.html', {'form': form})
     else:
-        form = AddProductForm(instance=product)
-
-    return render(request, 'products/product-add.html', {'form': form})
+        return redirect('home')
 
 
 def product_delete(request, pk):
-    product = get_object_or_404(Product, pk=pk)
-    product.delete()
-    return redirect('home')
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = get_object_or_404(Product, pk=pk)
+        product.delete()
+        return redirect('home')
+    else:
+        return redirect('home')
